VJA_imagegallery/requirements.txt
Pillow>=8.3.2
django-crispy-forms>=1.12.0
django-imagekit>=4.0.2
Django>=3.2.6
django-cleanup>=5.2.0
